IOF-343 testing: Split log testing code out into new file.
This method was in iof_test_local for historic reasons only, so
split it out into a new file to prepare for moving the code to
CaRT.

There should be no functional change with this, however I have
had to rework some of the includes and how classes are accessed.

Reviewed-on: https://review.hpdd.intel.com/33932
Tested-by: Jenkins
