+++ test/common_methods.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright (C) 2016-2018 Intel Corporation
+# Copyright (C) 2016-2019 Intel Corporation
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -74,58 +74,6 @@ try:
     COLORAMA = True
 except ImportError:
     COLORAMA = False
-
-# Use a global variable here so show_line can remember previously reported
-# error lines.
-shown_logs = set()
-
-def show_line(line, sev, msg):
-    """Output a log line in gcc error format"""
-
-    # Only report each individual line once.
-
-    log = "{}:{}:1: {}: {} '{}'".format(line.filename,
-                                        line.lineno,
-                                        sev,
-                                        msg,
-                                        line.get_anon_msg())
-    if log in shown_logs:
-        return
-    print(log)
-    shown_logs.add(log)
-
-def show_bug(line, bug_id):
-    """Mark output with a known bug"""
-    show_line(line, 'error', 'Known bug {}'.format(bug_id))
-
-class hwm_counter():
-    """Class to track integer values, with high-water mark"""
-
-    # Used for memory allocation sizes currently.
-    def __init__(self):
-        self.__val = 0
-        self.__hwm = 0
-        self.__acount = 0
-        self.__fcount = 0
-
-    def __str__(self):
-        return "Total:{:,} HWM:{:,} {} allocations, {} frees".\
-            format(self.__val,
-                   self.__hwm,
-                   self.__acount,
-                   self.__fcount)
-
-    def add(self, val):
-        """Add a value"""
-        self.__val += val
-        if self.__val > self.__hwm:
-            self.__hwm = self.__val
-        self.__acount += 1
-
-    def subtract(self, val):
-        """Subtract a value"""
-        self.__val -= val
-        self.__fcount += 1
 
 def import_list():
     """Return a array of imports
